Fix TestPolygon GeoJSON feature collection
